Rename okio-testing to okio-fakefilesystem and start publishing it
It's possible to use the fake filesystem in non-test situations, and
this new name is more descriptive.
