+++ okio-fakefilesystem/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=okio-fakefilesystem
+POM_NAME=Okio Fake Filesystem
