--- okio-testing/gradle.properties
@@ -1,2 +1,0 @@
-POM_ARTIFACT_ID=okio-testing
-POM_NAME=Okio Testing
