+++ okio/src/commonTest/kotlin/okio/ForwardingFilesystemTest.kt
@@ -17,6 +17,7 @@ package okio
 
 import kotlinx.datetime.Clock
 import okio.Path.Companion.toPath
+import okio.fakefilesystem.FakeFilesystem
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
