+++ settings.gradle
@@ -1,7 +1,7 @@
 rootProject.name = 'okio-parent'
 
 include ':okio'
-include ':okio-testing'
+include ':okio-fakefilesystem'
 include ':okio:jvm:japicmp'
 include ':okio:jvm:jmh'
 include ':samples'
